+++ tests/testthat/test-plot.R
@@ -76,14 +76,14 @@ test_that("First plot from plot_map renders as before", {
 
 #' @importFrom rlang .data
 test_that("calendar heatmap plot renders as before", {
+  local_mocked_bindings(
+    clean_fun = function(x) "en_US.UTF-8"
+  )
   vdiffr::expect_doppelganger(
     "A calendar heatmap plot",
-    withr::with_locale(
-      new = c("LC_TIME" = "en_US"),
-      plot_by_day_per_year_calendar_heatmap(
-        larger_input_for_testing,
-        display_mode = "dark"
-      )
+    plot_by_day_per_year_calendar_heatmap(
+      larger_input_for_testing,
+      display_mode = "dark"
     )
   )
 })