+++ NEWS.md
@@ -1,3 +1,8 @@
+# medacoPlot 1.1.0
+
+- language is a choice now
+- translations are available for english and german
+
 # medacoPlot 1.0.0
 
 - first stable version
@@ -6,7 +11,7 @@
 
 # medacoPlot 0.3.0
 
-- major redesign of the app 
+- major redesign of the app
 - renamed a lot of functions -> breaking change!
 - added data view for all plots
 - added initial darkmode support that still needs some improvements